Move ProtobufSimpleWriter to the internals of ProtobufWriter.
